+++ src/api/eval.rs
@@ -226,7 +226,7 @@ impl Engine {
             return Ok(Dynamic::UNIT);
         }
 
-        let lib = &[ast.lib()];
+        let lib = &[ast.as_ref()];
         self.eval_global_statements(scope, mods, &mut state, statements, lib, level)
     }
 }