+++ src/types/scope.rs
@@ -301,22 +301,21 @@ impl<'a> Scope<'a> {
     #[inline]
     #[must_use]
     pub fn contains(&self, name: &str) -> bool {
-        self.names
-            .iter()
-            .rev() // Always search a Scope in reverse order
-            .any(|(key, _)| name == key.as_ref())
+        self.names.iter().any(|(key, _)| name == key.as_ref())
     }
     /// Find an entry in the [`Scope`], starting from the last.
     #[inline]
     #[must_use]
     pub(crate) fn get_index(&self, name: &str) -> Option<(usize, AccessMode)> {
+        let len = self.len();
+
         self.names
             .iter()
             .rev() // Always search a Scope in reverse order
             .enumerate()
-            .find_map(|(index, (key, _))| {
+            .find_map(|(i, (key, _))| {
                 if name == key.as_ref() {
-                    let index = self.len() - 1 - index;
+                    let index = len - 1 - i;
                     Some((index, self.values[index].access_mode()))
                 } else {
                     None
@@ -338,16 +337,14 @@ impl<'a> Scope<'a> {
     #[inline]
     #[must_use]
     pub fn get_value<T: Variant + Clone>(&self, name: &str) -> Option<T> {
+        let len = self.len();
+
         self.names
             .iter()
             .rev()
             .enumerate()
             .find(|(_, (key, _))| name == key.as_ref())
-            .and_then(|(index, _)| {
-                self.values[self.len() - 1 - index]
-                    .flatten_clone()
-                    .try_cast()
-            })
+            .and_then(|(index, _)| self.values[len - 1 - index].flatten_clone().try_cast())
     }
     /// Check if the named entry in the [`Scope`] is constant.
     ///
@@ -522,14 +519,14 @@ impl<'a> Scope<'a> {
     #[inline]
     #[must_use]
     pub fn clone_visible(&self) -> Self {
+        let len = self.len();
+
         self.names.iter().rev().enumerate().fold(
             Self::new(),
             |mut entries, (index, (name, alias))| {
                 if !entries.names.iter().any(|(key, _)| key == name) {
                     entries.names.push((name.clone(), alias.clone()));
-                    entries
-                        .values
-                        .push(self.values[self.len() - 1 - index].clone());
+                    entries.values.push(self.values[len - 1 - index].clone());
                 }
                 entries
             },